+++ SandboxiePlus/QSbieAPI/Sandboxie/BoxBorder.cpp
@@ -21,6 +21,7 @@
 #include "../SbieAPI.h"
 
 #include <wtypes.h>
+#include <string>
 
 typedef HRESULT(*P_DwmIsCompositionEnabled)(BOOL *enabled);
 typedef HRESULT(*P_DwmGetWindowAttribute)(HWND hWnd, DWORD dwAttribute, void *pvAttribute, DWORD cbAttribute);
@@ -46,6 +47,13 @@ struct SBoxBorder
 	BOOL IsBorderVisible;
 	HWND BorderWnd;
 	HBRUSH BorderBrush;
+	HFONT LabelFont;
+	std::wstring BoxName;
+	int LabelWidth;
+	int LabelHeight;
+	int LabelPadding;
+	int LabelMode; // 0=disabled, -1=outside (above border), 1=inside (below border line)
+	RECT LabelRect;
 
 	int ThumbWidth;
 	int ThumbHeight;
@@ -61,6 +69,31 @@ void WINAPI CBoxBorder__TimerProc(HWND hwnd, UINT uMsg, UINT_PTR dwTimerID, DWOR
 {
 	CBoxBorder* This = (CBoxBorder*)GetWindowLongPtr(hwnd, 0);
 	This->TimerProc();
+}
+
+LRESULT CALLBACK CBoxBorder__WndProc(HWND hwnd, UINT uMsg, WPARAM wParam, LPARAM lParam)
+{
+	CBoxBorder* This = (CBoxBorder*)GetWindowLongPtr(hwnd, 0);
+	if (uMsg == WM_PAINT && This) {
+		PAINTSTRUCT ps;
+		HDC hdc = BeginPaint(hwnd, &ps);
+		SBoxBorder* m = This->GetBorderData();
+		if (m && m->LabelMode != 0 && !m->BoxName.empty() && m->LabelFont) {
+			HFONT hOldFont = (HFONT)SelectObject(hdc, m->LabelFont);
+			SetBkMode(hdc, TRANSPARENT);
+			// Use contrasting text color based on border color luminance
+			int r = GetRValue(m->BorderColor);
+			int g = GetGValue(m->BorderColor);
+			int b = GetBValue(m->BorderColor);
+			int luminance = (r * 299 + g * 587 + b * 114) / 1000;
+			SetTextColor(hdc, luminance > 128 ? RGB(0, 0, 0) : RGB(255, 255, 255));
+			DrawTextW(hdc, m->BoxName.c_str(), -1, &m->LabelRect, DT_CENTER | DT_VCENTER | DT_SINGLELINE);
+			SelectObject(hdc, hOldFont);
+		}
+		EndPaint(hwnd, &ps);
+		return 0;
+	}
+	return DefWindowProc(hwnd, uMsg, wParam, lParam);
 }
 
 DWORD WINAPI CBoxBorder__ThreadFunc(LPVOID lpParam)
@@ -103,6 +136,13 @@ void CBoxBorder::ThreadFunc()
 	m->IsBorderVisible = FALSE;
 	m->BorderWnd = NULL;
 	m->BorderBrush = NULL;
+	m->LabelFont = NULL;
+	m->BoxName.clear();
+	m->LabelWidth = 0;
+	m->LabelHeight = 0;
+	m->LabelPadding = 8;
+	m->LabelMode = -1; // Default to outside (above border)
+	memset(&m->LabelRect, 0, sizeof(RECT));
 
 	m->ThumbWidth = GetSystemMetrics(SM_CXHTHUMB);
 	m->ThumbHeight = GetSystemMetrics(SM_CYVTHUMB);
@@ -120,7 +160,7 @@ void CBoxBorder::ThreadFunc()
 	WNDCLASSEXW wc;
 	wc.cbSize = sizeof(WNDCLASSEX);
 	wc.style = CS_HREDRAW | CS_VREDRAW | CS_DBLCLKS | CS_GLOBALCLASS;
-	wc.lpfnWndProc = ::DefWindowProc;
+	wc.lpfnWndProc = CBoxBorder__WndProc;
 	wc.cbClsExtra = 0;
 	wc.cbWndExtra = sizeof(ULONG_PTR);
 	wc.hInstance = NULL; // (HINSTANCE)::GetModuleHandle(NULL);
@@ -156,10 +196,15 @@ void CBoxBorder::ThreadFunc()
 
 	KillTimer(m->BorderWnd, m->dwTimerId);
 
-	if (m->BorderWnd) 
+	if (m->BorderWnd)
 	{
 		DestroyWindow(m->BorderWnd);
 		m->BorderWnd = NULL;
+	}
+
+	if (m->LabelFont) {
+		DeleteObject(m->LabelFont);
+		m->LabelFont = NULL;
 	}
 }
 
@@ -186,7 +231,10 @@ void CBoxBorder::TimerProc()
 	{
 		m->pCurrentBox = pProcessBox.data();
 		if(!m->pCurrentBox)
+		{
 			m->BorderMode = 0;
+			m->BoxName.clear();
+		}
 		else
 		{
 			m->BorderMode = 1;
@@ -231,6 +279,21 @@ void CBoxBorder::TimerProc()
 					{
 						m->BorderAlpha = 192; // 75% opacity by default (192/255)
 					}
+
+					// Parse label mode (5th parameter) - no=disabled, in=inside, out=outside
+					if (BorderCfg.count() >= 5)
+					{
+						if(BorderCfg.at(4) == "no")
+							m->LabelMode = 0;
+						else if(BorderCfg.at(4) == "out")
+							m->LabelMode = -1;
+						else // "in" or default
+							m->LabelMode = 1;
+					}
+					else
+					{
+						m->LabelMode = 1; // Default to inside the border
+					}
 				}
 			}
 
@@ -242,6 +305,31 @@ void CBoxBorder::TimerProc()
 
 			// Apply alpha transparency setting
 			SetLayeredWindowAttributes(m->BorderWnd, 0, m->BorderAlpha, LWA_ALPHA);
+
+			// Store sandbox name and calculate label dimensions
+			m->BoxName = pProcessBox->GetName().toStdWString();
+
+			// Create font for label (bold, sized to fit in the border)
+			if (m->LabelFont) {
+				DeleteObject(m->LabelFont);
+				m->LabelFont = NULL;
+			}
+			int fontHeight = m->BorderWidth + 8; // Slightly larger than border width
+			m->LabelFont = CreateFontW(fontHeight, 0, 0, 0, FW_BOLD, FALSE, FALSE, FALSE,
+				DEFAULT_CHARSET, OUT_DEFAULT_PRECIS, CLIP_DEFAULT_PRECIS,
+				CLEARTYPE_QUALITY, DEFAULT_PITCH | FF_SWISS, L"Segoe UI");
+
+			// Calculate text dimensions
+			if (m->LabelFont) {
+				HDC hdc = GetDC(m->BorderWnd);
+				HFONT hOldFont = (HFONT)SelectObject(hdc, m->LabelFont);
+				SIZE textSize;
+				GetTextExtentPoint32W(hdc, m->BoxName.c_str(), (int)m->BoxName.length(), &textSize);
+				SelectObject(hdc, hOldFont);
+				ReleaseDC(m->BorderWnd, hdc);
+				m->LabelWidth = textSize.cx + m->LabelPadding * 2;
+				m->LabelHeight = textSize.cy + 4;
+			}
 		}
 	}
 
@@ -336,27 +424,130 @@ void CBoxBorder::TimerProc()
 			ah -= 1;
 
 
-		POINT Points[10];
-		int PointCount = 0;
+		HRGN hrgn = NULL;
+		int windowOffsetY = 0;
+		int windowHeight = ah;
 
-#define ADD_POINT(xx,yy) \
-		Points[PointCount].x = (xx);    \
-		Points[PointCount].y = (yy);    \
-		PointCount++;
+		if (m->LabelMode == 0)
+		{
+			// Mode 0: No label - simple rectangular border frame
+			POINT OuterPoints[5];
+			OuterPoints[0] = { 0, 0 };
+			OuterPoints[1] = { aw, 0 };
+			OuterPoints[2] = { aw, ah };
+			OuterPoints[3] = { 0, ah };
+			OuterPoints[4] = { 0, 0 };
 
-#define ADD_SQUARE(_w,_h,_b) \
-		ADD_POINT(0 + _b, 0 + _b); \
-		ADD_POINT(_w - _b, 0 + _b); \
-		ADD_POINT(_w - _b, _h - _b); \
-		ADD_POINT(0 + _b, _h - _b); \
-		ADD_POINT(0 + _b, 0 + _b);
+			HRGN hrgnOuter = CreatePolygonRgn(OuterPoints, 5, WINDING);
 
-		ADD_SQUARE(aw, ah, 0);
-		ADD_SQUARE(aw, ah, m->BorderWidth);
+			POINT InnerPoints[4];
+			InnerPoints[0] = { m->BorderWidth, m->BorderWidth };
+			InnerPoints[1] = { aw - m->BorderWidth, m->BorderWidth };
+			InnerPoints[2] = { aw - m->BorderWidth, ah - m->BorderWidth };
+			InnerPoints[3] = { m->BorderWidth, ah - m->BorderWidth };
 
-		HRGN hrgn = CreatePolygonRgn(Points, PointCount, ALTERNATE);
+			HRGN hrgnInner = CreatePolygonRgn(InnerPoints, 4, WINDING);
+			CombineRgn(hrgnOuter, hrgnOuter, hrgnInner, RGN_DIFF);
+			DeleteObject(hrgnInner);
+
+			hrgn = hrgnOuter;
+			memset(&m->LabelRect, 0, sizeof(RECT));
+		}
+		else if (m->LabelMode == -1)
+		{
+			// Mode -1: Label outside (above border)
+			int labelTabHeight = m->LabelHeight;
+			int labelTabWidth = m->LabelWidth;
+			int labelTabLeft = (aw - labelTabWidth) / 2;
+			int labelTabRight = labelTabLeft + labelTabWidth;
+
+			if (labelTabLeft < m->BorderWidth)
+				labelTabLeft = m->BorderWidth;
+			if (labelTabRight > aw - m->BorderWidth)
+				labelTabRight = aw - m->BorderWidth;
+
+			POINT OuterPoints[8];
+			int OuterCount = 0;
+			OuterPoints[OuterCount++] = { 0, labelTabHeight };
+			OuterPoints[OuterCount++] = { labelTabLeft, labelTabHeight };
+			OuterPoints[OuterCount++] = { labelTabLeft, 0 };
+			OuterPoints[OuterCount++] = { labelTabRight, 0 };
+			OuterPoints[OuterCount++] = { labelTabRight, labelTabHeight };
+			OuterPoints[OuterCount++] = { aw, labelTabHeight };
+			OuterPoints[OuterCount++] = { aw, ah + labelTabHeight };
+			OuterPoints[OuterCount++] = { 0, ah + labelTabHeight };
+
+			HRGN hrgnOuter = CreatePolygonRgn(OuterPoints, OuterCount, WINDING);
+
+			POINT InnerPoints[4];
+			InnerPoints[0] = { m->BorderWidth, labelTabHeight + m->BorderWidth };
+			InnerPoints[1] = { aw - m->BorderWidth, labelTabHeight + m->BorderWidth };
+			InnerPoints[2] = { aw - m->BorderWidth, ah + labelTabHeight - m->BorderWidth };
+			InnerPoints[3] = { m->BorderWidth, ah + labelTabHeight - m->BorderWidth };
+
+			HRGN hrgnInner = CreatePolygonRgn(InnerPoints, 4, WINDING);
+			CombineRgn(hrgnOuter, hrgnOuter, hrgnInner, RGN_DIFF);
+			DeleteObject(hrgnInner);
+
+			m->LabelRect.left = labelTabLeft;
+			m->LabelRect.top = 0;
+			m->LabelRect.right = labelTabRight;
+			m->LabelRect.bottom = labelTabHeight;
+
+			hrgn = hrgnOuter;
+			windowOffsetY = -labelTabHeight;
+			windowHeight = ah + labelTabHeight;
+		}
+		else // m->LabelMode == 1
+		{
+			// Mode 1: Label inside (thicker section at top center, extending into window area)
+			int labelTabHeight = m->LabelHeight;
+			int labelTabWidth = m->LabelWidth;
+			int labelTabLeft = (aw - labelTabWidth) / 2;
+			int labelTabRight = labelTabLeft + labelTabWidth;
+
+			if (labelTabLeft < m->BorderWidth)
+				labelTabLeft = m->BorderWidth;
+			if (labelTabRight > aw - m->BorderWidth)
+				labelTabRight = aw - m->BorderWidth;
+
+			// Outer rectangle is normal
+			POINT OuterPoints[5];
+			OuterPoints[0] = { 0, 0 };
+			OuterPoints[1] = { aw, 0 };
+			OuterPoints[2] = { aw, ah };
+			OuterPoints[3] = { 0, ah };
+			OuterPoints[4] = { 0, 0 };
+
+			HRGN hrgnOuter = CreatePolygonRgn(OuterPoints, 5, WINDING);
+
+			// Inner region: top at BorderWidth, but center section pushed down for label
+			POINT InnerPoints[8];
+			int InnerCount = 0;
+			InnerPoints[InnerCount++] = { m->BorderWidth, m->BorderWidth };
+			InnerPoints[InnerCount++] = { labelTabLeft, m->BorderWidth };
+			InnerPoints[InnerCount++] = { labelTabLeft, m->BorderWidth + labelTabHeight };
+			InnerPoints[InnerCount++] = { labelTabRight, m->BorderWidth + labelTabHeight };
+			InnerPoints[InnerCount++] = { labelTabRight, m->BorderWidth };
+			InnerPoints[InnerCount++] = { aw - m->BorderWidth, m->BorderWidth };
+			InnerPoints[InnerCount++] = { aw - m->BorderWidth, ah - m->BorderWidth };
+			InnerPoints[InnerCount++] = { m->BorderWidth, ah - m->BorderWidth };
+
+			HRGN hrgnInner = CreatePolygonRgn(InnerPoints, InnerCount, WINDING);
+			CombineRgn(hrgnOuter, hrgnOuter, hrgnInner, RGN_DIFF);
+			DeleteObject(hrgnInner);
+
+			m->LabelRect.left = labelTabLeft;
+			m->LabelRect.top = m->BorderWidth;
+			m->LabelRect.right = labelTabRight;
+			m->LabelRect.bottom = m->BorderWidth + labelTabHeight;
+
+			hrgn = hrgnOuter;
+		}
+
 		SetWindowRgn(m->BorderWnd, hrgn, TRUE);
-		SetWindowPos(m->BorderWnd, NULL, ax, ay, aw, ah, SWP_SHOWWINDOW | SWP_NOACTIVATE);
+		SetWindowPos(m->BorderWnd, NULL, ax, ay + windowOffsetY, aw, windowHeight, SWP_SHOWWINDOW | SWP_NOACTIVATE);
+		InvalidateRect(m->BorderWnd, NULL, TRUE);
 
 		m->IsBorderVisible = TRUE;
 	}