+++ CHANGELOG.md
@@ -11,7 +11,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [1.17.0 / 5.72.0] - 2026-01-??
 
 ### Added
-- Add icons to Sandboxie service applications #5160https://github.com/sandboxie-plus/Sandboxie/issues/5160#issuecomment-3706138019
+- Add icons to Sandboxie service applications [#5160](https://github.com/sandboxie-plus/Sandboxie/issues/5160#issuecomment-3706138019)
+- AddedShow name of sandbox in the border [#3746](https://github.com/sandboxie-plus/Sandboxie/issues/3746)
 
 ### Fixed
 - fixed failure to mount registry when starting an UWP application sandboxed